sk_ipoint_t and sk_isize_t should use integer members (#67)

+++ include/c/sk_types.h
@@ -175,7 +175,7 @@ typedef enum {
     TRANSLATE_SK_MATRIX44_TYPE_MASK = 0x01,
     SCALE_SK_MATRIX44_TYPE_MASK = 0x02,
     AFFINE_SK_MATRIX44_TYPE_MASK = 0x04,
-    PERSPECTIVE_SK_MATRIX44_TYPE_MASK = 0x08 
+    PERSPECTIVE_SK_MATRIX44_TYPE_MASK = 0x08
 } sk_matrix44_type_mask_t;
 
 /**
@@ -293,8 +293,8 @@ typedef struct {
 } sk_point3_t;
 
 typedef struct {
-    float   x;
-    float   y;
+    int32_t   x;
+    int32_t   y;
 } sk_ipoint_t;
 
 typedef struct {
@@ -303,8 +303,8 @@ typedef struct {
 } sk_size_t;
 
 typedef struct {
-    float   w;
-    float   h;
+    int32_t   w;
+    int32_t   h;
 } sk_isize_t;
 
 typedef struct {
@@ -570,7 +570,7 @@ typedef enum {
     INVERTED_SK_PATH_EFFECT_TRIM_MODE,
 } sk_path_effect_trim_mode_t;
 
-typedef struct sk_path_effect_t sk_path_effect_t;  
+typedef struct sk_path_effect_t sk_path_effect_t;
 
 typedef enum {
     BUTT_SK_STROKE_CAP,
@@ -891,8 +891,8 @@ typedef enum {
     PAETH_SK_PNGENCODER_FILTER_FLAGS = 0x80,
     ALL_SK_PNGENCODER_FILTER_FLAGS   = NONE_SK_PNGENCODER_FILTER_FLAGS |
                                        SUB_SK_PNGENCODER_FILTER_FLAGS |
-                                       UP_SK_PNGENCODER_FILTER_FLAGS | 
-                                       AVG_SK_PNGENCODER_FILTER_FLAGS | 
+                                       UP_SK_PNGENCODER_FILTER_FLAGS |
+                                       AVG_SK_PNGENCODER_FILTER_FLAGS |
                                        PAETH_SK_PNGENCODER_FILTER_FLAGS,
 } sk_pngencoder_filterflags_t;
